+++ inventory_management_system_api/services/system.py
@@ -7,7 +7,9 @@ from typing import Annotated, Optional
 
 from fastapi import Depends
 
-from inventory_management_system_api.core.exceptions import MissingRecordError
+from inventory_management_system_api.core.config import config
+from inventory_management_system_api.core.exceptions import MissingRecordError, ChildElementsExistError
+from inventory_management_system_api.core.object_storage_api_client import ObjectStorageAPIClient
 from inventory_management_system_api.models.system import SystemIn, SystemOut
 from inventory_management_system_api.repositories.system import SystemRepo
 from inventory_management_system_api.schemas.breadcrumbs import BreadcrumbsGetSchema
@@ -99,10 +101,19 @@ class SystemService:
 
         return self._system_repository.update(system_id, SystemIn(**{**stored_system.model_dump(), **update_data}))
 
-    def delete(self, system_id: str) -> None:
+    def delete(self, system_id: str, access_token: Optional[str] = None) -> None:
         """
         Delete a system by its ID
 
         :param system_id: ID of the system to delete
+        :param access_token: The JWT access token to use for auth with the Object Storage API if object storage enabled.
         """
-        return self._system_repository.delete(system_id)
+        if self._system_repository.has_child_elements(system_id):
+            raise ChildElementsExistError(f"System with ID {system_id} has child elements and cannot be deleted")
+
+        # First, attempt to delete any attachments and/or images that might be associated with this system.
+        if config.object_storage.enabled:
+            ObjectStorageAPIClient.delete_attachments(system_id, access_token)
+            ObjectStorageAPIClient.delete_images(system_id, access_token)
+
+        self._system_repository.delete(system_id)
